CET-12088 Refactore the code
