+++ src/components/ReportsPage/HeatmapPage/HeatmapFiltersSelect.tsx
@@ -22,7 +22,7 @@ import {
   Select,
 } from '@material-ui/core';
 import { Clear } from '@material-ui/icons';
-import React from 'react';
+import React, { useCallback, useMemo } from 'react';
 
 export interface ModalSelectItem {
   value: string | number;
@@ -30,42 +30,52 @@ export interface ModalSelectItem {
 }
 
 interface ModalSelectProps<T> {
+  configKey: T;
   name: string;
-  value: T[];
-  onChange: (value: T[]) => void;
-  onReset: () => void;
-  options: ModalSelectItem[];
+  value: string[];
+  onChange: (type: T, value: string[]) => void;
+  onReset: (type: T) => void;
+  options: string[];
 }
 
 export const ModalSelect = <T,>({
+  configKey,
   name,
   value,
   onChange,
   onReset,
   options,
 }: ModalSelectProps<T>) => {
+  const mappedOptions = useMemo(() => {
+    return options.map(option => (
+      <MenuItem key={`ScorecardOption-domain-${option}`} value={option}>
+        {option}
+      </MenuItem>
+    ));
+  }, [options]);
+
+  const changeHandler = useCallback(
+    (event: React.ChangeEvent<{ name?: string; value: unknown }>) => {
+      onChange(configKey, event.target.value as string[]);
+    },
+    [configKey, onChange],
+  );
+
+  const resetHandler = useCallback(() => {
+    onReset(configKey);
+  }, [onReset, configKey]);
+
   return (
     <Box display="flex" flexDirection="row">
       <FormControl variant="standard" fullWidth>
         <InputLabel>{name}</InputLabel>
-        <Select
-          multiple
-          value={value}
-          onChange={event => onChange(event.target.value as T[])}
-        >
-          {options.map(option => (
-            <MenuItem
-              key={`ScorecardOption-domain-${option.label}`}
-              value={option.value}
-            >
-              {option.label}
-            </MenuItem>
-          ))}
+        <Select multiple value={value} onChange={changeHandler}>
+          {mappedOptions}
         </Select>
       </FormControl>
       {value.length > 0 && (
         <Button
-          onClick={onReset}
+          onClick={resetHandler}
           variant="text"
           aria-label={`Clear ${name.toLocaleLowerCase()} filter`}
           title={`Clear ${name.toLocaleLowerCase()} filter`}
