+++ src/components/ReportsPage/HeatmapPage/HeatmapTable.tsx
@@ -13,30 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import React from 'react';
-import { EmptyState, Link } from '@backstage/core-components';
+import React, { useCallback, useMemo } from 'react';
+import { EmptyState } from '@backstage/core-components';
 import {
   useCortexBirdseye,
   BirdsEyeReportTable,
-  FilterType as BirdsEyeFilterType,
   TeamResponse,
-  DomainsHierarchyResponse as BirdsEyeDomainsHierarchyResponse,
   Domain,
-  Scorecard as BirdsEyeScorecard,
-  ScorecardDetailsScore,
   HierarchyResponse,
   StringIndexable,
   TeamDetails,
-  RuleOutcome as BirdsEyeRuleOutcome,
-  DomainTreeNode,
   Filters,
+  BirdsEyeReportTableProps,
 } from '@cortexapps/birdseye';
 import {
   DomainHierarchiesResponse,
-  DomainHierarchyNode,
-  EntityFilter,
-  FilterType,
-  RuleOutcome,
   Scorecard,
   ScorecardLadder,
   ScorecardServiceScore,
@@ -53,203 +44,14 @@ import {
   entityComponentRef,
 } from '../../../utils/ComponentUtils';
 import { parseEntityRef } from '@backstage/catalog-model';
-
-const convertToBirdsEyeFilterType = (
-  filterType: Exclude<EntityFilter['type'], FilterType.CQL_FILTER>,
-): Exclude<BirdsEyeFilterType, BirdsEyeFilterType.CQL_FILTER> => {
-  switch (filterType) {
-    case FilterType.SERVICE_FILTER:
-      return BirdsEyeFilterType.SERVICE_FILTER;
-    case FilterType.DOMAIN_FILTER:
-      return BirdsEyeFilterType.DOMAIN_FILTER;
-    case FilterType.RESOURCE_FILTER:
-      return BirdsEyeFilterType.RESOURCE_FILTER;
-    case FilterType.TEAM_FILTER:
-      return BirdsEyeFilterType.TEAM_FILTER;
-  }
-};
-
-const convertToBirdsEyeFilter = (
-  filter: Scorecard['filter'],
-): BirdsEyeScorecard['filter'] => {
-  return !filter
-    ? undefined
-    : filter?.type === FilterType.COMPOUND_FILTER
-    ? {
-        ...filter,
-        type: 'COMPOUND_FILTER',
-        cqlFilter: filter.cqlFilter
-          ? {
-              ...filter.cqlFilter,
-              type: BirdsEyeFilterType.CQL_FILTER,
-            }
-          : undefined,
-      }
-    : filter?.type === FilterType.CQL_FILTER
-    ? {
-        ...filter,
-        type: BirdsEyeFilterType.CQL_FILTER,
-      }
-    : {
-        ...filter,
-        type: convertToBirdsEyeFilterType(filter.type),
-      };
-};
-
-const convertToBirdsEyeScorecard = (
-  scorecard: Scorecard,
-  ladder: ScorecardLadder | undefined,
-): BirdsEyeScorecard => {
-  const rules: BirdsEyeScorecard['rules'] = scorecard.rules.map(rule => ({
-    ...rule,
-    cqlVersion: '', // TODO
-    dateCreated: rule.dateCreated ?? '',
-    filter: convertToBirdsEyeFilter(rule.filter),
-    id: rule.id.toString(),
-  }));
-
-  return {
-    ...scorecard,
-    exemptions: {
-      autoApprove: false,
-      enabled: false,
-    },
-    isDraft: false,
-    notifications: {
-      enabled: false,
-    },
-    filter: convertToBirdsEyeFilter(scorecard.filter),
-    id: scorecard.id.toString(),
-    rules,
-    ladder,
-  };
-};
-
-const convertToBirdsEyeRuleOutcome = (
-  rules: RuleOutcome[],
-): BirdsEyeRuleOutcome[] => {
-  return rules.map((rule: RuleOutcome): BirdsEyeRuleOutcome => {
-    const mappedRule = {
-      ...rule.rule,
-      cqlVersion: '',
-      dateCreated: rule.rule.dateCreated ?? '',
-      filter: undefined,
-      id: rule.rule.id.toString(),
-    };
-
-    return {
-      ...rule,
-      rule: mappedRule,
-    };
-  });
-};
-
-const convertToBirdsEyeScores = (
-  scores: ScorecardServiceScore[],
-  catalog: HomepageEntity[],
-): ScorecardDetailsScore[] => {
-  return scores.map((score: ScorecardServiceScore): ScorecardDetailsScore => {
-    const entity = catalog.find(item => item.id === score.serviceId);
-    return {
-      score: {
-        score: score.score,
-        scorePercentage: score.scorePercentage,
-        totalPossibleScore: score.totalPossibleScore,
-      },
-      entity: {
-        cid: '', // TODO
-        entityGroups: {
-          all: score.tags,
-          defined: score.tags,
-        },
-        entityOwners: [], // TODO
-        id: score.serviceId.toString(),
-        name: entity?.name ?? '',
-        ownerGroups: score.teams,
-        tag: score.componentRef,
-        type: entity?.type ?? '',
-        description: score.description,
-        icon: entity?.icon,
-      },
-      evaluation: {
-        rules: convertToBirdsEyeRuleOutcome(score.rules),
-        ladderLevels: score.ladderLevels,
-        lastUpdated: '',
-      },
-    };
-  });
-};
-
-const convertToBirdsEyeDomainNodeMetadata = (
-  item: DomainHierarchyNode,
-): DomainTreeNode => {
-  return {
-    ...item,
-    node: {
-      ...item.node,
-      description: item.node.description ?? undefined,
-      id: item.node.id.toString(),
-    },
-    orderedChildren: item.orderedChildren.map(
-      convertToBirdsEyeDomainNodeMetadata,
-    ),
-  };
-};
-
-const convertToBirdsEyeDomainHierarchy = (
-  hierarchies?: DomainHierarchiesResponse,
-): BirdsEyeDomainsHierarchyResponse | undefined => {
-  return {
-    ...hierarchies,
-    orderedTree:
-      hierarchies?.orderedTree.map(convertToBirdsEyeDomainNodeMetadata) ?? [],
-  };
-};
-
-const getScoreColorClassName = (points: number): string => {
-  if (points > 0.9) {
-    return 'success1';
-  } else if (points > 0.8) {
-    return 'success2';
-  } else if (points > 0.7) {
-    return 'success3';
-  } else if (points > 0.6) {
-    return 'warning1';
-  } else if (points > 0.5) {
-    return 'warning2';
-  } else if (points > 0.4) {
-    return 'warning3';
-  } else if (points > 0.3) {
-    return 'danger1';
-  } else if (points > 0.2) {
-    return 'danger2';
-  }
-
-  return 'danger3';
-};
-
-const getCellColorBackground = (value?: number): string => {
-  if (value === null || value === undefined) {
-    return 'danger3';
-  }
-
-  if (value > 0.9) {
-    return 'success1';
-  } else if (value > 0.49) {
-    return 'warning2';
-  } else {
-    return 'danger3';
-  }
-};
-
-const BirdsEyeAnchorAdapter = (
-  props: React.DetailedHTMLProps<
-    React.AnchorHTMLAttributes<HTMLAnchorElement>,
-    HTMLAnchorElement
-  >,
-) => {
-  return <Link to={props.href ?? ''}>{props.children}</Link>;
-};
+import {
+  BirdsEyeAnchorAdapter,
+  convertToBirdsEyeDomainHierarchy,
+  convertToBirdsEyeScorecard,
+  convertToBirdsEyeScores,
+  getCellColorBackground,
+  getScoreColorClassName,
+} from './HeatmapUtils';
 
 interface HeatmapTableProps {
   allDomains: Domain[];
@@ -283,17 +85,32 @@ export const HeatmapTable: React.FC<HeatmapTableProps> = ({
   teamHierarchy,
 }) => {
   const colorStyles = useColorCellStyles();
-  const mappedScorecard = convertToBirdsEyeScorecard(scorecard, ladder);
-  const mappedScores = convertToBirdsEyeScores(scores, catalog);
-  const mappedDomainHierarchies =
-    convertToBirdsEyeDomainHierarchy(domainHierarchy);
-  const mappedDomainAncestryMapKeys = mapKeys(domainAncestryMap, key =>
-    key.toString(),
-  );
-  const mappedDomainAncestryMap = mapValues(
-    mappedDomainAncestryMapKeys,
-    values => values.map(value => value.toString()),
-  );
+
+  const {
+    mappedDomainAncestryMap,
+    mappedDomainHierarchies,
+    mappedScorecard,
+    mappedScores,
+  } = useMemo(() => {
+    const mappedScorecard = convertToBirdsEyeScorecard(scorecard, ladder);
+    const mappedScores = convertToBirdsEyeScores(scores, catalog);
+    const mappedDomainHierarchies =
+      convertToBirdsEyeDomainHierarchy(domainHierarchy);
+    const mappedDomainAncestryMapKeys = mapKeys(domainAncestryMap, key =>
+      key.toString(),
+    );
+    const mappedDomainAncestryMap = mapValues(
+      mappedDomainAncestryMapKeys,
+      values => values.map(value => value.toString()),
+    );
+
+    return {
+      mappedScorecard,
+      mappedScores,
+      mappedDomainHierarchies,
+      mappedDomainAncestryMap,
+    };
+  }, [catalog, domainAncestryMap, domainHierarchy, ladder, scorecard, scores]);
 
   const {
     tableData,
@@ -325,6 +142,54 @@ export const HeatmapTable: React.FC<HeatmapTableProps> = ({
     scorecardServiceDetailsRouteRef,
   );
 
+  const cellColorClassNameHandler = useCallback(
+    (points: number) => {
+      return `${colorStyles.root} ${
+        colorStyles[getCellColorBackground(points)]
+      }`;
+    },
+    [colorStyles],
+  );
+
+  const scoreColorClassNameHandler = useCallback(
+    (points: number) => {
+      return `${colorStyles.root} ${
+        colorStyles[getScoreColorClassName(points)]
+      }`;
+    },
+    [colorStyles],
+  );
+
+  const scorecardEntityUrlHandler = useCallback(
+    (
+      scorecardEntity: Parameters<
+        BirdsEyeReportTableProps['getScorecardEntityUrl']
+      >[0],
+    ) => {
+      const entityId = Number.parseInt(scorecardEntity.entityId);
+      const entity = catalog.find(entity => entityId === entity.id);
+      const codeTag = entity?.codeTag;
+      if (!codeTag) {
+        return '';
+      }
+
+      const componentRef = entityComponentRef(entitiesByTag, codeTag);
+
+      const entityName = parseEntityRef(
+        componentRef,
+        defaultComponentRefContext,
+      );
+
+      const entityUrl = scorecardServiceDetailsRef({
+        scorecardId: `${mappedScorecard.id}`,
+        ...entityName,
+      });
+
+      return entityUrl;
+    },
+    [catalog, entitiesByTag, mappedScorecard.id, scorecardServiceDetailsRef],
+  );
+
   return (
     <Grid container direction={'column'}>
       <Grid item style={{ marginTop: '20px' }}>
@@ -352,34 +217,9 @@ export const HeatmapTable: React.FC<HeatmapTableProps> = ({
             <EmptyState title="Select a Scorecard" missing="data" />
           }
           filters={filters}
-          getCellColorClassName={points =>
-            `${colorStyles.root} ${colorStyles[getCellColorBackground(points)]}`
-          }
-          getScoreColorClassName={points =>
-            `${colorStyles.root} ${colorStyles[getScoreColorClassName(points)]}`
-          }
-          getScorecardEntityUrl={scorecardEntity => {
-            const entityId = Number.parseInt(scorecardEntity.entityId);
-            const entity = catalog.find(entity => entityId === entity.id);
-            const codeTag = entity?.codeTag;
-            if (!codeTag) {
-              return '';
-            }
-
-            const componentRef = entityComponentRef(entitiesByTag, codeTag);
-
-            const entityName = parseEntityRef(
-              componentRef,
-              defaultComponentRefContext,
-            );
-
-            const entityUrl = scorecardServiceDetailsRef({
-              scorecardId: `${mappedScorecard.id}`,
-              ...entityName,
-            });
-
-            return entityUrl;
-          }}
+          getCellColorClassName={cellColorClassNameHandler}
+          getScoreColorClassName={scoreColorClassNameHandler}
+          getScorecardEntityUrl={scorecardEntityUrlHandler}
           setFilters={setFilters}
         />
       </Grid>
